// src/types/next-auth.d.ts
import { DefaultSession } from "next-auth";

export type UserRole = "admin" | "warehouse" | "sales";

declare module "next-auth" {
  interface Session {
    user: {
      id: string;
      role: UserRole;
    } & DefaultSession["user"];
  }

  interface User {
    role: UserRole;
  }
}

declare module "next-auth/jwt" {
  interface JWT {
    id: string;
    role: UserRole;
  }
}